[Bugfix] Fixed multiple changes bug
Fixed case where there is multiple replacements in a diff. Now
is_content_in_framework will return all ranges that have have an
intersection. i.e. if there are two replacements in the same framework
function then a two tuples will be returned, where each tuple will
be paired with a framework fucntion reference, in this case the
framework function reference will be the same.

This way we can get all the changes that we need to make to a
particular framework function at once.