[task] Added base script for parsing logs
Added some skeleton code that will count the number of additions
and deltions made with a found bugfix for a given file. The main
function will return a triple for the given filename of the format:
	$filename $num_additions $num_deltions

Other information is available if needed such as the author, and the
date the commit was made.