[task] Added git metrics to add commit length
Added script git metrics to loop through the data dump csv from commit
guru and parse the data for the commit hashes. Using each of the commit
hashes it gets the length of the commit messages. From here it filters all
the rows in the original data dump that don't have a valid commit hash so
that the number of valid rows is equal to the size of the commit lengths
list i.e. there is a commit length per commit hash. From here it appends
the  new commit length data onto every row of data and writes out a new
csv containing the previous information and the new commit length
column.