[Refactor][Bugfix] Corrected to count commit lines
Modified the git command so that it returns only 1 commit
instead of all of them. Modified the counting of lines so
that it is counting lines always instead of counting
characters sometimes.

In the process of changing the way lines were counted
it was easy to also include the the number of words that
are contained in a commit as a new, similar metric as a
commit message may have multiple lines but not
necessarily have a lot of content to match.