[task] Update script to use parse csv's in dir
Updated script to allow for the parsing of single files as well
as all .csv files in a given directory. The output is then written
to a file that that is specified and if none is, a generic filename
merged_results will be used.

A producer consumer model is empolyed where the multiple threads
open and parse each of the files separtely and then post the processed
data into a queue. A consumer process then takes items from the queue
and writes them to a single file to imporce speed.