[task] Added porject identifier
The scipt takes as input a list of repository url's and uses
the github api to search for the word android. This  is first
searched in the description and then the readme and lastly the
code. The list of urls can be input via a command line option
or they can be found in a .csv file that is provided.

As output a list of projects is output and for each project there
is a colomn represneting if the the project is android or not.